chore(app/): switch to app router
* moved manifest and removed Meta component
