+++ components/contentful-image.js
@@ -1,3 +1,5 @@
+"use client";
+
 import Image from "next/image";
 
 const contentfulLoader = ({ src, width, quality }) => {
