+++ components/divisor/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   FunctionComponent,
   SVGAttributes,
