+++ lib/api.js
@@ -190,7 +190,7 @@ export async function getWelcomeForHome(preview) {
   };
 }
 
-export async function getPageAndMorePages(slug, preview) {
+export async function getPageAndMorePages(slug, preview = false) {
   const entry = await fetchGraphQL(
     `query {
       pageCollection(where: { slug: "${slug}" }, preview: ${
