--- pages/_app.tsx
@@ -1,9 +1,0 @@
-import type { AppProps } from "next/app";
-
-import "../styles/index.css";
-
-const MyApp = ({ Component, pageProps }: AppProps) => {
-  return <Component {...pageProps} />;
-};
-
-export default MyApp;
