--- pages/_document.tsx
@@ -1,15 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from "next/document";
-
-export default class MyDocument extends Document {
-  render() {
-    return (
-      <Html lang="en">
-        <Head />
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
